+++ CHANGELOG.md
@@ -19,8 +19,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 - Few panics that can be triggered by an invalid input.
-- `.` was parsed as `0` by `Steam::parse_number`. Now it's an error.
+- `.` was parsed as `0` by `Stream::parse_number`. Now it's an error.
 - Endless loop when a ClosePath segment was followed by a number.
+- `Stream::parse_integer` error when a number is at the end of the stream.
+- `Stream::parse_integer` will return an error on an integer overflow now.
 
 ## [0.2.1] - 2017-02-01
 ### Fixed